+++ groups/api.py
@@ -5,7 +5,7 @@ from django.db.models import Prefetch
 from rest_framework.parsers import MultiPartParser
 from rest_framework.viewsets import ModelViewSet
 
-from api.v2.mixins import CORSMixin
+from api.v2.mixins import METHODS, CORSMixin
 from api.v2.schema import OpenAPISchema
 
 from . import models, serializers
@@ -18,7 +18,6 @@ class GroupViewSet(CORSMixin, ModelViewSet):
     * list: List groups.
     * read: View a certain group.
     * create: Create a new group.
-    * update: Update the given group.
     * patch: Edit the given group.
     * delete: Delete the given group.
     """
@@ -32,6 +31,7 @@ class GroupViewSet(CORSMixin, ModelViewSet):
     )
     serializer_class = serializers.GroupSerializer
     parser_classes = (MultiPartParser,)
+    http_method_names = METHODS
 
 
 class MemberViewSet(CORSMixin, ModelViewSet):
@@ -41,7 +41,6 @@ class MemberViewSet(CORSMixin, ModelViewSet):
     * list: List members.
     * read: View a certain member.
     * create: Create a new member.
-    * update: Update the given member.
     * patch: Edit the given member.
     * delete: Delete the given member.
     """
@@ -53,6 +52,7 @@ class MemberViewSet(CORSMixin, ModelViewSet):
     )
     serializer_class = serializers.MemberSerializer
     parser_classes = (MultiPartParser,)
+    http_method_names = METHODS
 
 
 __all__ = ['GroupViewSet', 'MemberViewSet']