+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 v0.9.0
 ^^^^^^
 
+* Added new API endpoints
+* Disabled ``PUT`` operations
 * Dropped support for Python 3.7
 
 v0.8.5